Filtered topic option (#20)
* added USE_FILTERED_SCAN option

* Bump tiiuae/fog-ros-baseimage to dp-4266_humble_upgrade

* Fix errors in build

* Change baseimage tag from dp-4266_humble_upgrade to v2.0.0

---------

+++ sdk/src/rplidar_driver.cpp
@@ -1783,10 +1783,10 @@ u_result RPlidarDriverImplCommon::grabScanData(rplidar_response_measurement_node
 
     switch (_dataEvt.wait(timeout))
     {
-    case rp::hal::Event::EVENT_TIMEOUT:
+    case static_cast<unsigned long>(rp::hal::Event::EVENT_TIMEOUT):
         count = 0;
         return RESULT_OPERATION_TIMEOUT;
-    case rp::hal::Event::EVENT_OK:
+    case static_cast<unsigned long>(rp::hal::Event::EVENT_OK):
         {
             if(_cached_scan_node_hq_count == 0) return RESULT_OPERATION_TIMEOUT; //consider as timeout
 
@@ -1812,10 +1812,10 @@ u_result RPlidarDriverImplCommon::grabScanDataHq(rplidar_response_measurement_no
 {
     switch (_dataEvt.wait(timeout))
     {
-    case rp::hal::Event::EVENT_TIMEOUT:
+    case static_cast<unsigned long>(rp::hal::Event::EVENT_TIMEOUT):
         count = 0;
         return RESULT_OPERATION_TIMEOUT;
-    case rp::hal::Event::EVENT_OK:
+    case static_cast<unsigned long>(rp::hal::Event::EVENT_OK):
     {
         if (_cached_scan_node_hq_count == 0) return RESULT_OPERATION_TIMEOUT; //consider as timeout
